requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
requests==2.31.0
beautifulsoup4==4.12.2
python-multipart==0.0.6
jinja2==3.1.2
nltk==3.8.1
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.1.4
python-dotenv==1.0.0
aiofiles==23.2.1
httpx==0.25.2
trafilatura==2.0.0
lxml==5.4.0
lxml_html_clean==0.4.2
openai>=1.0.0
redis==5.0.1 